+++ ydb/mvp/core/mvp_test_runtime.h
@@ -2,6 +2,16 @@
 
 #include <ydb/library/actors/testlib/test_runtime.h>
 
+namespace {
+template <typename HttpType>
+void EatWholeString(TIntrusivePtr<HttpType>& request, const TString& data) {
+    request->EnsureEnoughSpaceAvailable(data.size());
+    auto size = std::min(request->Avail(), data.size());
+    memcpy(request->Pos(), data.data(), size);
+    request->Advance(size);
+}
+}
+
 class TMvpTestRuntime : public NActors::TTestActorRuntimeBase {
     using NActors::TTestActorRuntimeBase::TTestActorRuntimeBase;
