+++ ydb/mvp/oidc_proxy/ut/ya.make
@@ -4,6 +4,7 @@ SIZE(SMALL)
 
 SRCS(
     oidc_proxy_ut.cpp
+    oidc_token_exchange_ut.cpp
     openid_connect.cpp
 )
